requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
Django==2.2.2
django-debug-toolbar==1.11
ecdsa==0.13.2
entrypoints==0.3
flake8==3.7.7
future==0.17.1
idna==2.8
importlib-metadata==0.18
isort==4.3.20
jws==0.1.3
mccabe==0.6.1
more-itertools==7.0.0
Naked==0.1.31
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.2
pyflakes==2.1.1
pyparsing==2.4.0
pytest==4.6.3
python-decouple==3.1
python-jose==3.0.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.1
